Initialize Foreign Exchange (FX) Data

This function initializes and refreshes the local foreign exchange (FX) data directory by downloading the necessary data files from the specified remote server. The function allows different approaches for data management, including incremental updates, full reinitialization, and local refreshes.

Usage

fx_init(
  ...,
  fxsource = "ecb",
  verbose = TRUE,
  approach = c("incremental", "fresh", "local_refresh", "remove")
)

Arguments

...

Additional arguments (not currently used).

fxsource

Character string specifying the source of FX data. Currently, only "ecb" is supported.

verbose

Logical indicating whether to print progress messages. Defaults to TRUE.

approach

Character string specifying how the function should handle existing data. Must be one of "incremental" (default), "fresh", "local_refresh", "remove". See details.

Details

This function ensures that the local FX data store is up to date by:

Checking available data ranges on the remote server.

Identifying missing data files for download.

Removing outdated local files.

Loading data into a DuckDB database for efficient querying.

Currently, the function only supports the European Central Bank ("ecb") as the FX data source.

The approach to updating data is determined by the approach parameter. This can be set to four different options:

"incremental": Downloads missing data while keeping existing files.

"fresh": Deletes all local data and downloads everything anew.

"local_refresh": Rebuilds the local database without downloading new data.

"remove": Deletes all local data and exits without downloading new data.